requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==4.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Markdown==3.7
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
PyMuPDF==1.25.3
PyMySQL==1.1.1
python-decouple==3.8
reportlab==4.3.1
sqlparse==0.5.3
